Upgrade to Storybook 8.2
