--- .babelrc.js
@@ -1,6 +1,0 @@
-// this configuration is used by jest, not for actual building the addon
-
-module.exports = {
-    targets: "defaults",
-    presets: ["@babel/preset-env", "@babel/preset-typescript"],
-}
