+++ package.json
@@ -43,7 +43,6 @@
     "dist/**/*"
   ],
   "scripts": {
-    "clean": "rimraf ./dist",
     "start": "concurrently \"yarn storybook --quiet\" \"yarn build:dist --watch\"",
     "storybook": "storybook dev -p 6006",
     "test": "vitest",
@@ -53,32 +52,29 @@
     "prepublish": "yarn clean && yarn build:dist",
     "release": "auto shipit --base-branch=main"
   },
+  "dependencies": {
+    "@storybook/icons": "^1.2.10"
+  },
   "devDependencies": {
-    "@babel/preset-env": "^7.24.0",
-    "@babel/preset-typescript": "^7.23.3",
-    "@chromatic-com/storybook": "^1.2.1",
-    "@storybook/addon-essentials": "^8.0.0",
-    "@storybook/addon-interactions": "^8.0.0",
-    "@storybook/addon-links": "^8.0.0",
-    "@storybook/addon-webpack5-compiler-swc": "^1.0.0",
-    "@storybook/components": "^8.0.0",
-    "@storybook/core-events": "^8.0.0",
-    "@storybook/icons": "^1.2.9",
-    "@storybook/manager-api": "^8.0.0",
-    "@storybook/preview-api": "^8.0.0",
-    "@storybook/react-webpack5": "^8.0.0",
-    "@storybook/theming": "^8.0.0",
-    "@storybook/types": "^8.0.0",
+    "@chromatic-com/storybook": "^1.7.0",
+    "@storybook/addon-essentials": "^8.2.9",
+    "@storybook/addon-interactions": "^8.2.9",
+    "@storybook/addon-links": "^8.2.9",
+    "@storybook/react-vite": "^8.2.9",
     "@types/node": "^18.19.0",
+    "@vitejs/plugin-react": "^4.3.1",
     "auto": "^11.1.1",
     "concurrently": "^5.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "rimraf": "^3.0.2",
-    "storybook": "^8.0.0",
+    "storybook": "^8.2.9",
     "tsup": "^8.0.1",
     "typescript": "^4.9.4",
+    "vite": "^5.4.2",
     "vitest": "^2.0.5"
+  },
+  "peerDependencies": {
+    "storybook": "^8.2.0"
   },
   "bundler": {
     "exportEntries": [
