+++ src/manager.ts
@@ -1,4 +1,4 @@
-import { addons, types } from "@storybook/manager-api"
+import { addons, types } from "storybook/internal/manager-api"
 
 import { ADDON_ID, TOOL_ID } from "./constants"
 import { PseudoStateTool } from "./manager/PseudoStateTool"
