+++ src/stories/ShadowRootWithPart.css
