+++ src/stories/ShadowRootWithPart.jsx
